WTEL-900

+++ src/app/endpoint.go
@@ -2,7 +2,7 @@ package app
 
 import "github.com/webitel/acr/src/model"
 
-func (a *App) GetDistinctDevices(domainId int64, request []*model.EndpointsRequest) ([]*model.EndpointsResponse, error) {
+func (a *App) GetDistinctDevices(domainId int, request []*model.EndpointsRequest) ([]*model.EndpointsResponse, error) {
 	if result := <-a.Store.Endpoint().GetDistinctDevices(domainId, request); result.Err != nil {
 		return nil, result.Err
 	} else {
